refactor: getMaxApplicableDiscount 활인관련 로직으로 위치 이동
