+++ src/refactoring/models/cart/actions/calculate/discounts.ts
@@ -12,3 +12,11 @@ export const applyCouponDiscount = (
   }
   return amount * (1 - coupon.discountValue / 100);
 };
+
+export const getMaxApplicableDiscount = (item: CartItem) => {
+  const applicableDiscounts = item.product.discounts
+    .filter((discount) => item.quantity >= discount.quantity)
+    .map((discount) => discount.rate);
+
+  return applicableDiscounts.length > 0 ? Math.max(...applicableDiscounts) : 0;
+};
