+++ src/refactoring/models/cart/actions/calculate/index.ts
@@ -1,1 +1,4 @@
 export * from "./calculateCartTotal";
+// Q. 테스트코드 구조상 추가시켰는데 애초에 잘못 구성된걸까
+export * from "./itemCalculator";
+export * from "./discounts";
